Changed the style to reflect the solution in the GitHub issue
